style: display doubles with at most 2 decimals
Reformats doubles to be displayed with the least amount of decimal places (convert to integer if possible) and at most 2 decimal places.

+++ src/main/java/cielsachen/mco1/model/coffee/EspressoRatio.java
@@ -23,6 +23,7 @@ public enum EspressoRatio {
 
     @Override
     public String toString() {
-        return this.coffeeBean + " " + Ingredient.COFFEE_BEANS.name + " : " + this.water + " " + Ingredient.WATER.name;
+        return (int) this.coffeeBean + " " + Ingredient.COFFEE_BEANS.name + " : " + (int) this.water + " "
+                + Ingredient.WATER.name;
     }
 }
